+++ src/setup.jl
@@ -55,6 +55,7 @@ export read_topography
 export salinity_from_conductivity
 export SA
 export scale_bar
+export set_teos
 export subset_amsr
 export T90_from_T48
 export T90_from_T68